TASK: Make sure e2e test fail if the Publish Result Dialog shows an error
Currently, the `#neos-PublishDialog-Acknowledge` id will be used both for the "ok" button in success, and for the "cancel" button in case of any unexpected errors.

To make the e2e tests fail fast we change the id in the error case to `Acknowledge-Error` so we dont accidentally click it.

see also: https://github.com/neos/neos-ui/pull/3769#pullrequestreview-2332466270