BUGFIX: Further improve Discard all conflict resolution when publishing
While the basic functionality is restored with 66c11d6d971f23ab477a1a8ce8b8a0dfdfa7fb40, it now comes to and odd state that is caused by `actionTypes.CONFLICTS_RESOLVED` which moves the sync phase back to `ONGOING`.
Due to that there are now two modals during the finishing phase, once the sync rotating modal and on top the confirmation "1 change(s) in site "foo" were sucessfully published to workspace"

To fix the message the state mashine has to be properly replaced in nested calls for actionTypes.STARTED, as otherwise the modal would say that things were published, not discarded.

And we also stop the syncing workflow when discarding takes over to avoid two modals.